JSSC-66  🔨 Refactor : deletemodal 로직 변경
