+++ src/Page/GroupPage/Post/GroupPost.tsx
@@ -1,6 +1,6 @@
 import { styled } from 'styled-components';
 import { useState } from 'react';
-import PostHeader from '../PostHeader';
+import PostHeader from './PostHeader';
 import PostContent from './PostContent';
 import { PostProps } from '../PostPropsType';
