+++ src/Page/GroupPage/Post/PostHeader.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { styled } from 'styled-components';
-import point3 from '../../image/Icon/3point_icon.svg';
+import point3 from '../../../image/Icon/3point_icon.svg';
 import { PostProps } from '../PostPropsType';
 import DeleteModal from '../../../Component/DeleteModal';
 import { deleteBoard } from '../../../API/Boards';
